JSTL 도입하여 JSP에서 자바를 전부 제거

+++ src/webapp/jsp/article/list.jsp
@@ -1,13 +1,6 @@
-<%@ page import="java.util.List"%>
-<%@ page import="java.util.Map"%>
-<%@ page import="com.sbs.exam.dto.Article"%>
+<%@ taglib prefix="c" uri ="http://java.sun.com/jsp/jstl/core"%>
 <%@ page language="java" contentType="text/html; charset=UTF-8" pageEncoding="UTF-8"%>
 
-<%
-	List<Article> articles = (List<Article>) request.getAttribute("articles");
-	int cPage = (int) request.getAttribute("page");
-	int totalPage = (int) request.getAttribute("totalPage");
-%>
 <!DOCTYPE html>
 <html>
 <head>
@@ -16,7 +9,6 @@
 </head>
 <body>
 	<h1>게시물 리스트</h1>
-
 	<%@ include file="../part/topBar.jspf"%>
 
 	<div>
@@ -35,22 +27,18 @@
 			</tr>
 		</thead>
 		<tbody>
-			<%
-				for ( Article article : articles) {
-			%>
+			<c:forEach items="${articles}" var="article">
 			<tr>
-				<td><%=article.id%></td>
-				<td><%=article.regDate%></td>
-				<td><a href="detail?id=<%=article.id%>"><%=article.title%></a></td>
+				<td>${article.id}</td>
+				<td>${article.regDate}</td>
+				<td><a href="detail?id=${article.id}">${article.title}</a></td>
 				<td>
-					<a href="doDelete?id=<%=article.id%>">삭제</a>
-					<a href="modify?id=<%=article.id%>">수정</a>
+					<a onclick="if ( confirm('삭제하시겠습니까?') == false ) return false;" href="doDelete?id=${article.id}">삭제</a>
+					<a href="modify?id=${article.id}">수정</a>
 				</td>
 			</tr>
+			</c:forEach>
 		</tbody>
-		<%
-			}
-		%>
 	</table>
 	<style type="text/css">
 	    .page > a.red {
@@ -58,26 +46,28 @@
 	    }
 	</style>
 	<div class="page">
-		<% if ( cPage >1 ) { %>
-			<a href="list?page=1">◀</a>
-		<% } %>
-		<%
-			int pageMenuSize = 10;
-			int from = cPage - pageMenuSize;
-			if ( from < 1) {
-				from = 1;
-			}
-			int end = cPage + 10;
-			if ( end > totalPage ) {
-				end = totalPage;
-			}
-			for (int i = from; i <= end; i++) {
-		%>
-			<a class="<%=cPage == i ? "red" : ""%>" href="list?page=<%=i%>"><%=i%></a>
-		<% } %>
+	    <c:set var="cPage" value="${page}" />
+	    <c:set var="totalPage" value="${totalPage}" />
+	    <c:set var="pageMenuSize" value="10" />
+        <c:set var="from" value="${cPage - pageMenuSize}" />
+
+        <c:if test="${cPage > 1}" >
+           <a href="list?page=1">◀</a>
+       </c:if>
+
+       <c:set var="start" value="${from < 1 ? 1 : from }" />
+       <c:set var="end" value="${cPage + 10}" />
+
+       <c:set var="end" value="${end > totalPage ? totalPage : end}"/>
+
+       <c:forEach var="i" begin="${start}" end="${end}" step="1">
+          <c:set var="aClassRed" value="${cPage == i ? 'red' : ''}"/>
+          <a class="${aClassRed}" href="list?page=${i}">${i}</a>
+      </c:forEach>
 	</div>
-	<% if ( cPage < totalPage ) { %>
-		<a href="list?page=<%=totalPage%>">▶</a>
-	<% } %>
+
+	<c:if test="${cPage > totalPage}" >
+        <a href="list?page=${totalPage}">▶</a>
+    </c:if>
 </body>
 </html>
